[FBS-133] - refactor: moving form to flowbuild server panel component
